function-app/requirements.txt
azure-functions==1.18.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
requests==2.31.0
python-dateutil==2.8.2